+++ R/intermodal_route.R
@@ -31,6 +31,7 @@ intermodal_route <- function(origin, destination, datetime = Sys.time(),
   # Checks
   .check_points(origin)
   .check_points(destination)
+  .check_input_rows(origin, destination)
   .check_datetime(datetime)
   .check_numeric_range(results, 1, 7)
   .check_numeric_range(transfers, -1, 6)